refactor: share logic
